Reportsコンポーネントをview部分とlogic部分で分離
