+++ app/javascript/components/Report/CommentUserIcon.jsx
