+++ app/javascript/components/Report/PracticeFilterDropdown.jsx
