+++ app/views/companies/reports/index.html.slim
@@ -14,5 +14,4 @@ header.page-header
 
 .page-body
   .container.is-md
-    div(data-vue="Reports" data-vue-company-id:number="#{@company.id}")
-    = react_component('Reports', companyId: @company.id)
+    = react_component('Report/CompanyReports', companyId: @company.id)
