+++ app/views/current_user/reports/index.html.slim
@@ -4,7 +4,7 @@
 
 .page-body
   .container.is-md
-    = react_component('Reports', userId: current_user.id)
+    = react_component('Report/UserReports', userId: current_user.id)
     - if current_user.reports.present?
       .form-actions
         ul.form-actions__items
