+++ app/views/practices/reports/index.html.slim
@@ -6,4 +6,4 @@
 
 .page-body
   .container.is-md
-    = react_component('Reports', practiceId: @practice.id)
+    = react_component('Report/PracticeReports', practiceId: @practice.id)
