+++ app/views/reports/index.html.slim
@@ -12,4 +12,4 @@ header.page-header
 .page-body
   .container.is-md
 
-= react_component('Reports', all: true, practices: current_user.practices)
+= react_component('Report/PracticesReports', practices: current_user.practices)
