+++ app/views/reports/unchecked/index.html.slim
@@ -11,4 +11,4 @@ header.page-header
 
 .page-body
   .container.is-md
-    = react_component('Reports', unchecked: true)
+    = react_component('Report/UncheckedReports')
