+++ app/views/users/reports/index.html.slim
@@ -6,4 +6,4 @@
 
 .page-body
   .container.is-md
-    = react_component('Reports', userId: @user.id, practices: @user.practices)
+    = react_component('Report/UserPracticesReports', userId: @user.id, practices: @user.practices)
